feat(components): Add components to separate forro and samba classes
